部署脚本

+++ BlogCloudServer/bin/bulid.sh
@@ -1,25 +1,34 @@
 #! /bin/bash
 
-rm -rf /opt/blog
+rm -rf /opt/blog/Blog-1.0.jar
+rm -rf /opt/blog/www
 if [[ ! -d "/opt/BlogCloudJersey" ]]; then
 	cd /opt
 	git clone git@github.com:zhanghaijun666/BlogCloudJersey.git
 else
 	cd /opt/BlogCloudJersey
+	git stash
+	git stash clear
 	git pull
 fi
 cd /opt/BlogCloudJersey/BlogCloudServer
 mvn clean package
 if [[ ! -d "/opt/blog" ]]; then
-	mkdir /opt/blog
+	mkdir -p /opt/blog
 fi
 cd /opt/blog
 ln -s /opt/BlogCloudJersey/BlogCloudServer/target/BlogCloudServer-1.0-SNAPSHOT.jar /opt/blog/Blog-1.0.jar
 ln -s /opt/BlogCloudJersey/BlogCloudClient/www /opt/blog/www
-cp -r /opt/BlogCloudJersey/BlogCloudServer/config /opt/blog/config
-cp /opt/BlogCloudJersey/BlogCloudServer/user.txt /opt/blog/user.txt
-cp -r /opt/BlogCloudJersey/BlogCloudServer/bin /opt/blog/bin
-chmod -R +x /opt/blog/bin/*.sh
+if [[ ! -d "/opt/blog/config" ]]; then
+	cp -r /opt/BlogCloudJersey/BlogCloudServer/config /opt/blog/config
+fi
+if [[ ! -f "/opt/blog/user.txt" ]]; then
+	cp /opt/BlogCloudJersey/BlogCloudServer/user.txt /opt/blog/user.txt
+fi
+if [[ ! -d "/opt/blog/bin" ]]; then
+	cp -r /opt/BlogCloudJersey/BlogCloudServer/bin /opt/blog/bin
+	chmod -R +x /opt/blog/bin/*.sh
+fi
 if [[ ! -d "/usr/lib/systemd/system" ]]; then
 	mkdir -p /usr/lib/systemd/system
 fi
